GLFW 3.2 bindings (dev!) except custom cursors.
Not tested even slightly. I think it compiles.

Note that the Vulkan stuff referring to MarshalAs LPStr is probably
fine, as the Vulkan functions talk in ASCII, but I bound a non-Vulkan
function in particular as byte* rather than using MarshalAs to go
straight to a string as the documentation explicitly says it's UTF8.

Vulkan-specific params are mostly IntPtr, as it's likely you'd prefer the
types of a separate Vulkan C# binding library.

Unlike other wrapped pointers, GlfwVulkanProcPtr has inner_ptr public as
you'll likely need to extract it to pass it on to something else.

The custom cursor stuff was not bound because it seems even more tedious
than doing all this was.

I remain amazingly unclear on what happens for 'int' return values from
P/Invoke. e.g. One func returns 'VkResult', which is a C enum, which is
equivalent to 'int' w.r.t. the whims of the C compiler. Nuts. Don't ask
me. C int -> C# 'int' is used elsewhere in the bindings, so I assume it
usually just magically works...

Please don't expect me to update this. I have no interest in the 32-bit
bindings, and am going to fork this privately and work with a much more
squished-down version of the ol' bindings as is my wont.